pyproject.toml
[project]
name = "swap-bot-py"
version = "0.1.0"
description = "A Python bot project managed by Poetry"
authors = [
    {name = "tychefi"}
]
readme = "README.md"
requires-python = ">3.9.1,<4.0"
dependencies = [
    "pyeoskit (>=1.1.12,<2.0.0)",
    "coincurve (>=21.0.0,<22.0.0)",
    "cryptography (>=45.0.6,<46.0.0)",
    "pycryptodome (>=3.23.0,<4.0.0)",
    "base58 (>=2.1.1,<3.0.0)",
    "pyflonkit @ git+https://github.com/fullon-labs/pyflonkit.git",
    "pyyaml (>=6.0.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^1.1.0"
pytest = "^8.4.1"

